functorial interface for Ocsipersist
New functorial interface allows for different column types in
Sqlite/PostgreSQL instead of storing everything as text/blobs/strings.
This way keys and values can be stored and retrieved without
necessarily marshalling/unmarshalling them. This allows for better
readability of the data in the DB for instance via the psql interactive
terminal, and for higher-performance organisation of the data.